# swagger.yaml
swagger: '2.0'
info:
  description: "Milla is mostly Qt Jira-specific orchestrator service for the OpenReq Qt trial that operates between Qt Jira and the OpenReq infrastructure. Milla has five main functionalities divided into three different controllers. <br>\r\n<b>Milla controller:</b><br>1) Milla fetches all or updated data from Qt Jira and stores in a caching manner the data in the Mallikas database service. <br>2) Milla orchestrates the communication to Mallikas database service so that a project, requirement etc. can be manipulated (CRUD manner) in Mallikas database service in OpenReq JSON format.<br>3) Milla orchestrates the data from Mallikas database service to Mulperi service in order to manage dependencies, carry out analyses etc. \n<br><b>Detection controller:</b><br>4) Milla calls the natural language processing services of OpenReq infrastructre that can detect from the requirements text hidden or implicit dependencies to other requirements.  <br>\r\n<b>Qt controller:</b><br>5) Milla provides the access point for the Qt Jira dependency browser in the Qt trial. <br>\r\n"
  version: '1.13'
  title: Milla
host: 'localhost:9203'
basePath: /
...
paths:
...
  /qtJiraUpdated:
    post:
      tags:
        - milla-controller
      summary: Import the updated issues from Qt Jira to Mallikas.
      description: '<b>Functionality</b>: Import the updated Jira issues of a project from Qt Jira and send the issues as OpenReq JSON requirements and dependencies to be cached in Mallikas database<br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: New or changed Jira issues are updated to Mallikas<br><b>Exception</b>: This is not full synchronization.<br><b>Parameter: </b><br>projectId: The id of the Qt Jira project (e.g., QTWB).'
      operationId: importUpdatedFromQtJiraUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: projectId
          description: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Success, all updated requirements and dependencies downloaded'
          schema:
            type: string
        '201':
          description: Created
        '400':
          description: 'Failure, ex. malformed JSON'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: 'Failure, ex. invalid URLs'
  /qtjira:
    post:
      tags:
        - milla-controller
      summary: Import a selected project from Qt Jira and store a cache of the project in Mallikas
      description: '<b>Functionality</b>: This is the full data import from Qt Jira. A selected project including its issues are fetched from Qt Jira. The data is converted to OpenReq JSON requirements and dependencies, and OpenReq JSON project is constructed.The project including its requirements and dependencies is cached to Mallikas database<br><b>Postcondition</b>: The selected project is cached in Mallikas to be managed in the OpenReq infrastructure. The same project name (id) is used in Qt Jira and Mallikas.<br><b>Note:</b> For update rather than full import, see "qtjiraUpdated", which is  more efficient for large projects.</b><br><b>Parameter: </b><br>projectId: The id of the project to be fetched from Qt Jira (e.g., QTWB).'
      operationId: importFromQtJiraUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: projectId
          description: projectId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: 'Success, all requirements and dependencies downloaded'
          schema:
            type: string
        '400':
          description: 'Failure, ex. malformed JSON'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: 'Failure, ex. invalid URLs'
  /requirementsByIds:
    post:
      tags:
        - milla-controller
      summary: Fetch a specified set of requirements in OpenReq JSON format.
      description: '<b>Functionality</b>: Fetch a specied requirements including their dependencies that are cached in Mallikas database in the OpenReq JSON format.<br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: An OpenReq JSON of the requirements and their dependencies is produced.<br><b>Parameter: </b><br>ids: ids as a String array, e.g. ["QTWB-1", "QTWB-2"] '
      operationId: getRequirementsByIdsUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: ids
          description: ids
          required: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /requirementsByParams:
    post:
      tags:
        - milla-controller
      summary: Fetch requirements based on various parameters posted in JSON format.
      description: '<b>Functionality</b>: Fetch requirements based on various parameters given in JSON format.<br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: An OpenReq JSON of the requirements and their dependencies is produced.<br><b>Parameter: </b><br>requestParams: Parameters object containing various parameters'
      operationId: getRequirementsByParamsUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: params
          description: params
          required: true
          schema:
            $ref: '#/definitions/RequestParams'
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /requirementsInProject:
    get:
      tags:
        - milla-controller
      summary: Fetch all requirements  of a project in OpenReq JSON format.
      description: '<b>Functionality</b>: Fetch all requirements of a project including their dependencies that are cached in Mallikas database in the OpenReq JSON format.<br><b>Precondition</b>: The project has been cached in Mallikas.<br><b>Postcondition</b>: An OpenReq JSON of the project including requirements and dependencies is produced.<br><b>Parameter: </b><br>projectId: The project id in Mallikas (e.g., QTWB).'
      operationId: getRequirementsInProjectUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: projectId
          in: query
          description: projectId
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /sendProjectToMulperi:
    post:
      tags:
        - milla-controller
      summary: Construct a transitive closure by sending all requirements of the selected project to Mulperi
      description: '<b>Functionality</b>: All requirements of a project are fetched from Mallikas database and sent to Mulperi to construct a transitive closure.<br><b>Precondition</b>: The project, including its requirements, has been cached in Mallikas.<br><b>Postcondition</b>: Mulperi has a transitive closure of each requirement for a project up to depth five.<br><br><b>Parameter: </b><br>projectId: The project id in Mallikas (e.g., QTWB).'
      operationId: sendProjectToMulperiUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: projectId
          in: query
          description: projectId
          required: true
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /sendProjectToMulperiWithParams:
    post:
      tags:
        - milla-controller
      summary: Construct a transitive closure by sending all requirements of the selected project to Mulperi
      description: '<b>Functionality</b>: All requirements of a project are fetched from Mallikas database and sent to Mulperi to construct a transitive closure.<br><b>Precondition</b>: The project, including its requirements, has been cached in Mallikas.<br><b>Postcondition</b>: Mulperi has a transitive closure of each requirement for a project up to depth five.<br><br><b>Parameters: </b><br>projectId: The project id in Mallikas (e.g., QTWB).<br>includeProposedDependencies: Whether to fetch dependencies with the status ''proposed'''
      operationId: sendProjectToMulperiWithParamsUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: projectId
          in: query
          description: projectId
          required: true
          type: string
        - name: includeProposedDependencies
          in: query
          description: includeProposedDependencies
          required: true
          type: boolean
      responses:
        '200':
          description: OK
          schema:
            type: object
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
...
definitions:
  RequestParams:
    type: object
    properties:
      created_at:
        type: string
        format: date-time
      includeProposed:
        type: boolean
      includeRejected:
        type: boolean
      maxDependencies:
        type: integer
        format: int32
      modified_at:
        type: string
        format: date-time
      projectId:
        type: string
      proposedOnly:
        type: boolean
      requirementIds:
        type: array
        items:
          type: string
      resolution:
        type: string
      scoreThreshold:
        type: number
        format: double
      status:
        type: string
      type:
        type: string